backend/requirements.txt
Flask>=2.0
python-dotenv>=0.20
pdfminer.six>=20201018
pillow>=9.0.0
pytesseract>=0.3.9
fpdf2>=2.7.0
pytest>=7.0
flask-cors>=4.0.0
dotenv
